CallbackList : fix bool() fonction
When testing if a CallbackList is empty, it could return a wrong result. That's because when removing a callback it is only set to null in the internal array of callback. But the array itself stays the same size and the bool operator check the size.
So now use a map with a counter. Callbacks are really removed from the map and the bool opertor returns the correct result.