Widget : add MouseDown event
I will mouse the click event to a new base widget Clickable. Only things like button will have the onclick event. So I need to add events onMouseDownand onMouseUp for all the other widgets.
The clickevent will have a specific behavior : the mouse down and up will have to be inside the widget and the onclick will be trigger on the up event.